Made Fourier Transformations work
Removed margins and padding for a seamless experience. Generates a fourier transformation by interpreting given 2D Coordinates in variable "test" as Complex numbers in form "a + bi". DFT (Discrete Fourier Transform) is applied and Frequency, Phase, and Amplitude are calculated from the Real and Imaginary components of the DFT Set.